refactor(server): extract CoreConfig from types.ts to config.ts
`app/types.ts` mixed boot-input config (`CoreConfig`) with host-primitive
types (`CoreApp`, hook types, ID re-exports). They're consumed by
different layers — `CoreConfig` is what callers pass INTO `createCoreApp`
(boot), while `CoreApp` is the live handle they get BACK (host). Pulling
them apart makes the boot/host distinction explicit.

- `CoreConfig` moves to `app/config.ts` (next to `ServerConfigLoader`
  and `LoadedConfig` — all boot-input types in one file).
- Update 3 import sites to read `CoreConfig` from `./config.js` (or
  `./app/config.js` for standalone.ts).
- `types.ts` is now pure host primitives (CoreApp + hook types +
  UserId/AgentId re-exports). Drops imports of Db, SessionValidator,
  WebhookClient, Layer, TraceCaptureTag — no longer needed there.

Tests: 30/30 pass across src/app and standalone.
