+++ packages/server/src/app/server.ts
@@ -29,7 +29,8 @@ import { appHandlers } from "./handlers/apps.handlers.js";
 
 import { WebhookClient } from "../adapters/webhook.js";
 
-import type { CoreConfig, CoreApp } from "./types.js";
+import type { CoreApp } from "./types.js";
+import type { CoreConfig } from "./config.js";
 import {
   AppHostTag,
   ConversationServiceTag,
