+++ packages/server/src/standalone.ts
@@ -16,7 +16,8 @@ import { makeEffectKysely } from "./db/effect-kysely-toolkit.js";
 import { WebhookClient } from "./adapters/webhook.js";
 import { WebhookContactService } from "./adapters/webhook-contact-service.js";
 import { WebhookSessionValidator } from "./identity/services/session-validator.js";
-import type { CoreApp, CoreConfig } from "./app/types.js";
+import type { CoreApp } from "./app/types.js";
+import type { CoreConfig } from "./app/config.js";
 import type { Database } from "./db/database.js";
 import type { Db } from "./db/client.js";
 import { PostgresDialect } from "./db/postgres-dialect.js";
